+++ packages/coinbase/src/index.ts
@@ -1,1 +1,1 @@
-export * from './CoinbaseConnector'
+export * from './coinbaseConnector'
