+++ packages/core/src/index.ts
@@ -1,8 +1,8 @@
 export * from './browserWalletConnector'
 export * from './utils'
 export * from './errors'
-export * from './useVueDapp'
 export * from './types'
+export * from './useVueDapp'
 
 import VueDappProvider from './VueDappProvider.vue'
 export { VueDappProvider }
