+++ packages/walletconnect/src/index.ts
@@ -1,1 +1,1 @@
-export * from './WalletConnectConnector'
+export * from './walletConnectConnector'
